rpc tests: Inline dependency used in one place
